+++ src/ColorTools/HslColor.cs
@@ -54,7 +54,7 @@ namespace ColorTools
             return ToRgb().ToColor();
         }
 
-        public Color ToColor(byte alpha)
+        public Color ToColor(double alpha)
         {
             return ToRgb().ToColor(alpha);
         }
